UGenInfo: load resources from the same ClassLoader
Rather than forcing use of the system class loader, we should load the
needed resources from the same class loader that was used to load this
very class that needs them.

This fixes a bug where the resources cannot be found if UGenInfo was
loaded by, e.g., a URLClassLoader.